[fea] Track context when parsing include files
Previously we would always parse files (including files referenced via
`include` statements) as if they were standalone FEA. This means that we
would only accept statements that were valid at the top-level FEA scope.

Some sources, however, use `include` statements inside of feature
blocks, where additional statements are permitted.

This patch keeps track of the type of the node in which an include
statement is encountered, and picks an entry point for parsing that file
based on this scope.

Currently we just support 'top level' and 'feature', which unblocks us;
we can expand on this as needed.