[ir] Instantiate missing intermediates when flattening
This fixes an issue where a composite glyph with an intermediate layer
would cause a crash during flattening if any of its components did not
have a layer at the same location.

With this patch, we will interpolate a new instance at the missing
location for the component, and use that.