[ir] Improve testing of instantiation
Specifically improving the test setup code, and adding basic tests for
the cases where a composite has a location missing in one of its
components, and vice-versa.

The second of these cases is not currently handled correctly, but the
test validates the current behaviour and demonstrates we at least don't
crash.