[ir] Don't propogate instances while flattening
In an earlier patch I was a bit too excited about ensuring that
composite glyphs were defined at the locations of their (nested)
components.

We need to do this during decomposition but not during flattening, since
flattening isn't converting components to contours, and components are
allowed to define different locations than the composites that reference
them.