[ir] Fix logic bug in decomposition
We were making sure that the composite glyph was defined at the
component glyph locations, but we were not actually using that updated
glyph when we then went to make sure each component was defined at the
appropriate locations.

There is a remaining issue with how we handle non-trivial affines, but
this patch at least fixes a crash in a number of sources.