[glyphs] Fix logic bug in category construction
A misplaced 'filter_map' was causing us to skip the (null) attributes of
the first glyph in a ligature sequence if that glyph had a name we could
not recognize. This caused us to treat the second glyph as the first,
and so miscategorize various glyphs as non-spacing marks, which was
causing various little diffs.