[crater] Fixup cache invalidation logic
Previously we would use the python dependencies (the output of `pip
freeze`) as well as the presence of any changes to ttx_diff since the previously
run commit to invalidate caches.

As ttx_diff is now a python package, and is installed from a local path,
it ins included in the output of pip freeze with a sha appeneded, and
this sha changes on each commit to the fontc repo.

The solution here is just to manually remove ttx_diff from the output of
pip freeze, since changes are still tracked independently.