fix notify-plugin workflow to use workflow_run instead of release event
Releases created by `softprops/action-gh-release` action (which we use in our release.yml workflow) don't trigger `release: [published]` events (because GitHub prevents GITHUB_TOKEN actions from triggering other workflows to avoid recursive loops). So when I released fontc-v0.6.0 yesterday the fontc-export-plugin repo was not notified automatically.

To trigger that particular we would have to manually create a Github Release, but we prefer instead to push a tag and let the CI do the rest.

To achieve the same we can instead use `workflow_run` to trigger the notify-plugin workflow when the release workflow completes, see:

https://docs.github.com/en/actions/reference/workflows-and-actions/events-that-trigger-workflows#workflow_run