[fontir] 'Fix' overflow in overlay_feature_variations
This was ported directly from the python,  and the python uses an
(arbitrary precision) python integer to track the rank. We were using a
u64 here, and there exists at least one font (playfair) where there end
up being more than 64 unique regions for feature variations.

This new solution tries to balance performance and correctness: we can
handle up to 511 unique regions, will panic in debug if this is
exceeded, and will log if we exceed in release.

I am not 100% confident in this approach, or in 511 as a reasonable
maximum, but it's a starting point.