[name] Refactor NameBuilder, skip names at the end
So in order to match ufo2ft, we want to avoid generating fallback names
if a name was declared in the source, even if it was empty and we will
skip it later.

This implements that by only clearing out the empty names at the end.

To make 'the end' slightly more clearly defined, this also includes a
minor refactor to NameBuilder to combine 'apply_default_fallbacks' and
'into_inner' into a single 'build' method. This ensures that empty
strings are always cleared at the end, and more cannot be added
afterwards.