Fixup sub/superscript metric fallback logic
If there are explicit values in a font for the 'Os2Subscript___'
metrics, we want to reuse those for the corresponding superscript
metrics.

This means that we need to query metrics in this routine, which ran into
borrowck issues due to the long-lived closure that had a mutable ref to
self, so I've removed that.